refactor: share SplitAndTrimCSV util for comma-separated lists
