+++ pkg/cmd/cli/install/install.go
@@ -37,6 +37,7 @@ import (
 	"github.com/vmware-tanzu/velero/pkg/install"
 	velerotypes "github.com/vmware-tanzu/velero/pkg/types"
 	"github.com/vmware-tanzu/velero/pkg/uploader"
+	"github.com/vmware-tanzu/velero/pkg/util"
 	kubeutil "github.com/vmware-tanzu/velero/pkg/util/kube"
 )
 
@@ -330,7 +331,7 @@ func (o *Options) AsVeleroOptions() (*install.VeleroOptions, error) {
 		Plugins:                          o.Plugins,
 		NoDefaultBackupLocation:          o.NoDefaultBackupLocation,
 		CACertData:                       caCertData,
-		Features:                         splitAndTrimCSV(o.Features),
+		Features:                         util.SplitAndTrimCSV(o.Features),
 		DefaultVolumesToFsBackup:         o.DefaultVolumesToFsBackup,
 		UploaderType:                     o.UploaderType,
 		DefaultSnapshotMoveData:          o.DefaultSnapshotMoveData,
@@ -597,19 +598,4 @@ func (o *Options) Validate(c *cobra.Command, args []string, f client.Factory) er
 	}
 
 	return nil
-}
-
-// splitAndTrimCSV splits a comma-separated string and trims whitespace so values
-// like "EnableCSI, EnableAPIGroupVersions" produce the intended feature names.
-func splitAndTrimCSV(s string) []string {
-	parts := strings.Split(s, ",")
-	out := make([]string, 0, len(parts))
-	for _, part := range parts {
-		item := strings.TrimSpace(part)
-		if item == "" {
-			continue
-		}
-		out = append(out, item)
-	}
-	return out
 }
